Reorganize beneficio y categoria modelos
