+++ api/Abstracciones/Modelos/Categoria.cs
@@ -1,4 +1,5 @@
-﻿using System.ComponentModel.DataAnnotations;
+﻿using System;
+using System.ComponentModel.DataAnnotations;
 
 namespace Abstracciones.Modelos
 {
@@ -22,5 +23,25 @@ namespace Abstracciones.Modelos
     {
         // Datos descriptivos opcionales
         
+    }
+
+    public class CategoriaEnUsoResponse : CategoriaResponse
+    {
+        public string Code { get; set; } = "CategoriaEnUso";
+        public string Message { get; set; } = "La categoría tiene beneficios asociados.";
+        public int Count { get; set; }
+    }
+
+    public class CategoriaEnUsoException : Exception
+    {
+        public Guid CategoriaId { get; }
+        public int BeneficiosCount { get; }
+
+        public CategoriaEnUsoException(Guid categoriaId, int beneficiosCount, string? message = null, Exception? innerException = null)
+            : base(message ?? "La categoría tiene beneficios asociados.", innerException)
+        {
+            CategoriaId = categoriaId;
+            BeneficiosCount = beneficiosCount;
+        }
     }
 }
